@@ -29,16 +29,16 @@
   },
   "packageManager": "yarn@3.2.4",
   "dependencies": {
-    "@emotion/react": "^11.11.0",
-    "@emotion/styled": "^11.11.0",
+    "@emotion/react": "^11.14.0",
+    "@emotion/styled": "^11.14.0",
     "airtable": "^0.12.2",
-    "next": "^13.4.19",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0"
+    "next": "^15.3.1",
+    "react": "^19.1.0",
+    "react-dom": "^19.1.0"
   },
   "devDependencies": {
-    "@biomejs/biome": "^1.8.3",
-    "@emotion/babel-plugin": "^11.11.0",
+    "@biomejs/biome": "^1.9.4",
+    "@emotion/babel-plugin": "^11.13.5",
     "@mui/icons-material": "^5.14.3",
     "@mui/material": "^5.14.5",
     "@storybook/addon-actions": "^7.0.6",
@@ -47,10 +47,10 @@
     "@storybook/addon-links": "^7.0.6",
     "@storybook/nextjs": "^7.0.6",
     "@styled/typescript-styled-plugin": "^1.0.1",
-    "@types/node": "^18.15.11",
-    "@types/react": "^18.2.47",
-    "@types/react-dom": "^18.2.18",
-    "@typescript-eslint/eslint-plugin": "^5.57.1",
+    "@types/node": "^22.15.3",
+    "@types/react": "^19.1.2",
+    "@types/react-dom": "^19.1.3",
+    "@typescript-eslint/eslint-plugin": "^8.31.1",
     "hamburger-react": "^2.5.0",
     "lint-staged": "^13.0.3",
     "postcss-styled-syntax": "^0.4.0",
@@ -67,6 +67,6 @@
     "stylelint-config-styled-components": "^0.1.1",
     "stylelint-order": "^6.0.4",
     "stylelint-prettier": "^5.0.0",
-    "typescript": "^5.4.2"
+    "typescript": "^5.8.3"
   }
 }
