@@ -5,9 +5,9 @@ import type { PageProps } from "@/types";
 import { css } from "@emotion/react";
 import type { NextPage } from "next";
 import Image from "next/image";
-import ThemeImage from "public/2025/images/keyvisual.png";
-import ETHTokyoLogo from "public/assets/ETHTokyoLogo.png";
 import type { FC } from "react";
+import ThemeImage from "/public/2025/images/keyvisual.png";
+import ETHTokyoLogo from "/public/assets/ETHTokyoLogo.png";
 
 // テクニカルグラフィック
 const TechGraphic: FC = () => {
