@@ -14,105 +14,105 @@ import { Inter } from "next/font/google";
 import Head from "next/head";
 import Image, { type StaticImageData } from "next/image";
 import Link from "next/link";
-import AbhimanyuPic from "public/2024/images/contributors/abhimanyunag.jpg";
-import AggrePic from "public/2024/images/contributors/aggre.jpg";
-import AllenPic from "public/2024/images/contributors/allenjoseph.jpg";
-import AnjaliPic from "public/2024/images/contributors/anjaliyoung.png";
-import AudreyPic from "public/2024/images/contributors/audrey.png";
-import AyaPic from "public/2024/images/contributors/ayamiyaguchi.png";
-import ChuazhengPic from "public/2024/images/contributors/chuazhenglong.jpeg";
-import DangerPic from "public/2024/images/contributors/dangerz.png";
-import DanielPic from "public/2024/images/contributors/danielhwang.png";
-import DaodudePic from "public/2024/images/contributors/daodude.jpg";
-import DeliPic from "public/2024/images/contributors/deligong.png";
-import EdmundPic from "public/2024/images/contributors/edmundedgar.jpg";
-import FrancescoPic from "public/2024/images/contributors/francescoandreoli.jpg";
-import FriscoPic from "public/2024/images/contributors/friscodanconia.jpeg";
-import GuruPic from "public/2024/images/contributors/guru.jpg";
-import JamesPic from "public/2024/images/contributors/jameshe.jpeg";
-import JehyukPic from "public/2024/images/contributors/jehyukjang.jpg";
-import JongwonPic from "public/2024/images/contributors/jongwonpark.jpg";
-import JosephPic from "public/2024/images/contributors/josephschiarizzi.jpg";
-import JoshPic from "public/2024/images/contributors/joshguha.png";
-import KolbyPic from "public/2024/images/contributors/kolbymorozliebl.jpg";
-import LeonaPic from "public/2024/images/contributors/leohio.jpeg";
-import LouiePic from "public/2024/images/contributors/louieoconnor.jpg";
-import MattPic from "public/2024/images/contributors/matthewliu.jpeg";
-import MelisaPic from "public/2024/images/contributors/melisakurtcan.jpg";
-import MoazzamPic from "public/2024/images/contributors/moazzamarif.png";
-import NabihaPic from "public/2024/images/contributors/nabihasheikh.png";
-import NaimPic from "public/2024/images/contributors/naimashhab.jpeg";
-import NikhilPic from "public/2024/images/contributors/nikhilmahana.jpg";
-import Nilspic from "public/2024/images/contributors/nilspihl.png";
-import OzgurPic from "public/2024/images/contributors/ozgurarmancyigit.jpg";
-import PeterPic from "public/2024/images/contributors/peterszilagyi.png";
-import PhilipPic from "public/2024/images/contributors/phillippieper.jpg";
-import PiyushPic from "public/2024/images/contributors/piyushgururani.jpeg";
-import QiangPic from "public/2024/images/contributors/qiangju.png";
-import QiPic from "public/2024/images/contributors/qizhou.png";
-import RenaPic from "public/2024/images/contributors/renaobrien.png";
-import RohitPic from "public/2024/images/contributors/rohitramesh.jpg";
-import RyoheiPic from "public/2024/images/contributors/ryokomiyama.jpg";
-import SammiPic from "public/2024/images/contributors/sammishu.jpg";
-import SamuelPic from "public/2024/images/contributors/samuelchong.png";
-import ShogoPic from "public/2024/images/contributors/shogoochiai.jpg";
-import SuhyeonPic from "public/2024/images/contributors/suhyeonlee.png";
-import SujithPic from "public/2024/images/contributors/sujithsomraaj.jpeg";
-import TakaPic from "public/2024/images/contributors/takaphaver.jpg";
-import TomaszPic from "public/2024/images/contributors/tomaszstanczak.png";
-import TomoPic from "public/2024/images/contributors/tomosaito.jpg";
-import VarunPic from "public/2024/images/contributors/varundoshi.jpg";
-import VitalikPic from "public/2024/images/contributors/vitalik.jpg";
-import VladPic from "public/2024/images/contributors/vladsvitanko.jpg";
-import YosuiPic from "public/2024/images/contributors/yosuiharasawa.jpg";
-import YukiPic from "public/2024/images/contributors/yukiyuminaga.jpeg";
-import YuroitakiPic from "public/2024/images/contributors/yuroitaki.jpeg";
-import topHero from "public/2024/images/hero_top.jpg";
-import seikaiha from "public/2024/images/seikaiha.png";
-import API3Logo from "public/2024/logo/API3Logo.svg";
-import AkindoLogo from "public/2024/logo/AkindoLogo.svg";
-import AvailLogo from "public/2024/logo/AvailLogo.svg";
-import BNBLogo from "public/2024/logo/BNBLogo.svg";
-import BitflyerLogo from "public/2024/logo/BitflyerLogo.svg";
-import BunzzLogo from "public/2024/logo/BunzzLogo.svg";
-import CabinetLogo from "public/2024/logo/CabinetLogo.png";
-import CentrumLogo from "public/2024/logo/CentrumLogo.svg";
-import CoinpostLogo from "public/2024/logo/CoinpostLogo.png";
-import CryptonomadsLogo from "public/2024/logo/CryptonomadsLogo.svg";
-import CryptopolitanLogo from "public/2024/logo/CryptopolitanLogo.svg";
-import DeFormLogo from "public/2024/logo/DeFormLogo.png";
-import DesciJapanLogo from "public/2024/logo/DesciJapanLogo.png";
-import DevconLogo from "public/2024/logo/DevconLogo.svg";
-import EJLogoFace from "public/2024/logo/EJLogoFace_dark.png";
-import ENSLogo from "public/2024/logo/ENSLogo.svg";
-import ESPLogo from "public/2024/logo/ESPLogo.svg";
-import ETHDiamondLogo from "public/2024/logo/ETHDiamondLogo.png";
-import ETHRiyadhLogo from "public/2024/logo/ETHRiyadhLogo.svg";
-import ETHTaipeiLogo from "public/2024/logo/ETHTaipeiLogo.png";
-import FractonLogo from "public/2024/logo/FractonLogo.png";
-import GMOLogo from "public/2024/logo/GMOLogo.png";
-import IntmaxLogo from "public/2024/logo/IntmaxLogo.svg";
-import LxDaoLogo from "public/2024/logo/LxDaoLogo.png";
-import MercoinLogo from "public/2024/logo/MercoinLogo.png";
-import MetamaskLogo from "public/2024/logo/MetamaskLogo.svg";
-import MoongateLogo from "public/2024/logo/MoongateLogo.svg";
-import NEOLogo from "public/2024/logo/NEO-X.svg";
-import NeroLogo from "public/2024/logo/NeroLogo.svg";
-import NewEconomyLogo from "public/2024/logo/NewEconomyLogo.svg";
-import RedbullLogo from "public/2024/logo/RedbullLogo.svg";
-import ScrollLogo from "public/2024/logo/ScrollLogo.svg";
-import ShibuyaLogo from "public/2024/logo/Shibuya.svg";
-import SolidityScanLogo from "public/2024/logo/SolidityScanLogo.svg";
-import StreamEthLogo from "public/2024/logo/StreamEthLogo.png";
-import TechfundLogo from "public/2024/logo/TechfundLogo.png";
-import UnchainLogo from "public/2024/logo/UnchainLogo.png";
-import ZeloLogo from "public/2024/logo/ZeloLogo.png";
-import FiveDollarVcLogo from "public/2024/logo/fivedollarvc.png";
-import ETHTokyoLogo from "public/assets/ETHTokyoLogo.png";
 import { useEffect, useState } from "react";
 import type { FC, ReactNode } from "react";
 import { GoInfo } from "react-icons/go";
 import { HiCalendarDays, HiOutlineMapPin } from "react-icons/hi2";
+import AbhimanyuPic from "/public/2024/images/contributors/abhimanyunag.jpg";
+import AggrePic from "/public/2024/images/contributors/aggre.jpg";
+import AllenPic from "/public/2024/images/contributors/allenjoseph.jpg";
+import AnjaliPic from "/public/2024/images/contributors/anjaliyoung.png";
+import AudreyPic from "/public/2024/images/contributors/audrey.png";
+import AyaPic from "/public/2024/images/contributors/ayamiyaguchi.png";
+import ChuazhengPic from "/public/2024/images/contributors/chuazhenglong.jpeg";
+import DangerPic from "/public/2024/images/contributors/dangerz.png";
+import DanielPic from "/public/2024/images/contributors/danielhwang.png";
+import DaodudePic from "/public/2024/images/contributors/daodude.jpg";
+import DeliPic from "/public/2024/images/contributors/deligong.png";
+import EdmundPic from "/public/2024/images/contributors/edmundedgar.jpg";
+import FrancescoPic from "/public/2024/images/contributors/francescoandreoli.jpg";
+import FriscoPic from "/public/2024/images/contributors/friscodanconia.jpeg";
+import GuruPic from "/public/2024/images/contributors/guru.jpg";
+import JamesPic from "/public/2024/images/contributors/jameshe.jpeg";
+import JehyukPic from "/public/2024/images/contributors/jehyukjang.jpg";
+import JongwonPic from "/public/2024/images/contributors/jongwonpark.jpg";
+import JosephPic from "/public/2024/images/contributors/josephschiarizzi.jpg";
+import JoshPic from "/public/2024/images/contributors/joshguha.png";
+import KolbyPic from "/public/2024/images/contributors/kolbymorozliebl.jpg";
+import LeonaPic from "/public/2024/images/contributors/leohio.jpeg";
+import LouiePic from "/public/2024/images/contributors/louieoconnor.jpg";
+import MattPic from "/public/2024/images/contributors/matthewliu.jpeg";
+import MelisaPic from "/public/2024/images/contributors/melisakurtcan.jpg";
+import MoazzamPic from "/public/2024/images/contributors/moazzamarif.png";
+import NabihaPic from "/public/2024/images/contributors/nabihasheikh.png";
+import NaimPic from "/public/2024/images/contributors/naimashhab.jpeg";
+import NikhilPic from "/public/2024/images/contributors/nikhilmahana.jpg";
+import Nilspic from "/public/2024/images/contributors/nilspihl.png";
+import OzgurPic from "/public/2024/images/contributors/ozgurarmancyigit.jpg";
+import PeterPic from "/public/2024/images/contributors/peterszilagyi.png";
+import PhilipPic from "/public/2024/images/contributors/phillippieper.jpg";
+import PiyushPic from "/public/2024/images/contributors/piyushgururani.jpeg";
+import QiangPic from "/public/2024/images/contributors/qiangju.png";
+import QiPic from "/public/2024/images/contributors/qizhou.png";
+import RenaPic from "/public/2024/images/contributors/renaobrien.png";
+import RohitPic from "/public/2024/images/contributors/rohitramesh.jpg";
+import RyoheiPic from "/public/2024/images/contributors/ryokomiyama.jpg";
+import SammiPic from "/public/2024/images/contributors/sammishu.jpg";
+import SamuelPic from "/public/2024/images/contributors/samuelchong.png";
+import ShogoPic from "/public/2024/images/contributors/shogoochiai.jpg";
+import SuhyeonPic from "/public/2024/images/contributors/suhyeonlee.png";
+import SujithPic from "/public/2024/images/contributors/sujithsomraaj.jpeg";
+import TakaPic from "/public/2024/images/contributors/takaphaver.jpg";
+import TomaszPic from "/public/2024/images/contributors/tomaszstanczak.png";
+import TomoPic from "/public/2024/images/contributors/tomosaito.jpg";
+import VarunPic from "/public/2024/images/contributors/varundoshi.jpg";
+import VitalikPic from "/public/2024/images/contributors/vitalik.jpg";
+import VladPic from "/public/2024/images/contributors/vladsvitanko.jpg";
+import YosuiPic from "/public/2024/images/contributors/yosuiharasawa.jpg";
+import YukiPic from "/public/2024/images/contributors/yukiyuminaga.jpeg";
+import YuroitakiPic from "/public/2024/images/contributors/yuroitaki.jpeg";
+import topHero from "/public/2024/images/hero_top.jpg";
+import seikaiha from "/public/2024/images/seikaiha.png";
+import API3Logo from "/public/2024/logo/API3Logo.svg";
+import AkindoLogo from "/public/2024/logo/AkindoLogo.svg";
+import AvailLogo from "/public/2024/logo/AvailLogo.svg";
+import BNBLogo from "/public/2024/logo/BNBLogo.svg";
+import BitflyerLogo from "/public/2024/logo/BitflyerLogo.svg";
+import BunzzLogo from "/public/2024/logo/BunzzLogo.svg";
+import CabinetLogo from "/public/2024/logo/CabinetLogo.png";
+import CentrumLogo from "/public/2024/logo/CentrumLogo.svg";
+import CoinpostLogo from "/public/2024/logo/CoinpostLogo.png";
+import CryptonomadsLogo from "/public/2024/logo/CryptonomadsLogo.svg";
+import CryptopolitanLogo from "/public/2024/logo/CryptopolitanLogo.svg";
+import DeFormLogo from "/public/2024/logo/DeFormLogo.png";
+import DesciJapanLogo from "/public/2024/logo/DesciJapanLogo.png";
+import DevconLogo from "/public/2024/logo/DevconLogo.svg";
+import EJLogoFace from "/public/2024/logo/EJLogoFace_dark.png";
+import ENSLogo from "/public/2024/logo/ENSLogo.svg";
+import ESPLogo from "/public/2024/logo/ESPLogo.svg";
+import ETHDiamondLogo from "/public/2024/logo/ETHDiamondLogo.png";
+import ETHRiyadhLogo from "/public/2024/logo/ETHRiyadhLogo.svg";
+import ETHTaipeiLogo from "/public/2024/logo/ETHTaipeiLogo.png";
+import FractonLogo from "/public/2024/logo/FractonLogo.png";
+import GMOLogo from "/public/2024/logo/GMOLogo.png";
+import IntmaxLogo from "/public/2024/logo/IntmaxLogo.svg";
+import LxDaoLogo from "/public/2024/logo/LxDaoLogo.png";
+import MercoinLogo from "/public/2024/logo/MercoinLogo.png";
+import MetamaskLogo from "/public/2024/logo/MetamaskLogo.svg";
+import MoongateLogo from "/public/2024/logo/MoongateLogo.svg";
+import NEOLogo from "/public/2024/logo/NEO-X.svg";
+import NeroLogo from "/public/2024/logo/NeroLogo.svg";
+import NewEconomyLogo from "/public/2024/logo/NewEconomyLogo.svg";
+import RedbullLogo from "/public/2024/logo/RedbullLogo.svg";
+import ScrollLogo from "/public/2024/logo/ScrollLogo.svg";
+import ShibuyaLogo from "/public/2024/logo/Shibuya.svg";
+import SolidityScanLogo from "/public/2024/logo/SolidityScanLogo.svg";
+import StreamEthLogo from "/public/2024/logo/StreamEthLogo.png";
+import TechfundLogo from "/public/2024/logo/TechfundLogo.png";
+import UnchainLogo from "/public/2024/logo/UnchainLogo.png";
+import ZeloLogo from "/public/2024/logo/ZeloLogo.png";
+import FiveDollarVcLogo from "/public/2024/logo/fivedollarvc.png";
+import ETHTokyoLogo from "/public/assets/ETHTokyoLogo.png";
 import judges from "./judges-2024.json";
 import mentors from "./mentors-2024.json";
 
